[encryption-service] Copy all content from /home. Fixes JB#46131
Change preparation scripts to copy all content from /home partition
(except lost+found) if there is enough space in /tmp. Otherwise just
/home/nemo is copied and warning is logged.

Remove marker file first instead of last so that preparation is not
started again on next boot even if /home could not be preserved. In the
current form encrypting again doesn't work anyway.

Introduces dependency on coreutils because many of those tools are used
in the script.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>