[unlock-agent] Implement sending of temporary key. Fixes JB#46019
Implement sending of temporary key if marker file exists. This temporary
key is used as passphrase and it is attempted only once. If it fails,
passphrase was not temporary key and thus it needs to be asked from the
user. This implemented only to handle unusual case of needing to boot a
device which has temporary key marker set and on UI side this currently
doesn't indicate this any other way than by disabling input. Once there
is some sort of busy indicator, this can behave better. Invalid
passphrase message is not shown if sending temporary key fails.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>