[sailfish-unlock-ui] Fix speaker button state. Fixes JB#46444
Speaker button state was not changed properly when it was pressed
because the state doesn't change immediately in Call class. This is
fixed by introducing EarpieceOn and SpeakerOn statuses that are signaled
back to UI. They change based on signal from ohmd.

Also add small descriptions for Call::Status values.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>