Move all password agent tasks to new Agent class.
Introduce new Agent class that handles what main.cpp used to and some of
PinUi agent tasks as well. This allows to send temporary passphrase
before UI is shown or compositor is registered.

[sailfish-unlock-ui] Send temporary key before showing ui. Fixes JB#46517

Make PinUi a child of Agent and move Compositor to be a subclass of
Agent. State changes from Compositor are exposed to PinUi. This code
requires newer sailfish-minui-devel so the dependency is also updated.
Leave out reading of unused values from ask file and store the read
values in Agent instance.

Because Compositor is now a subclass of Agent, no GUI is required
to do Compositor registration or other tasks. Thus creation of UI is
moved back to PinUi constructor and some checks are removed from
PinUI.

Remove old temporary passphrase logic since that is handled by Agent.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>