Handle ofono state better.
There was a bug that if you made a call but ended it immediately it
would actually leave emergency call mode and still make the call. Now it
will not attempt to call if user cancels the call before ofono's Dial is
called. Also calling is changed to better reflect the state that user
expects it to be so that cancel button is actually ending the call when
it is supposed to.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>