[unlock-ui] Fix restarting emergency call immediately. Fixes JB#48104
If bringing up modem was already in progress when the next call is
started, it will fail to org.ofono.Error.InProgress. Change it to just
return there and let the original handler to start the call.

Also reset timer on Call::Status::Initializing, so that it will not
leave emergency call view and end the call due to that.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>