[sailfish-device-encryption] Make emergency button behavior consistent again. JB#52739
The code was fighting on two places whether to show or hide the
emergency button in the emergency mode. The normal unlock UI
hides it on the emergency mode, exiting with keypad cancel button.

This used to work that way too until commit 81ba6032f132
which was mainly doing other things. Feels like just gotten wrong there.

In addition to being inconsistent, this was also rendered bad
due to black background baked into the the emergency button svg.